Completed localization
* Esp. for the log view output
* Removed the copyright from the log view's welcome message
* Added de.catkeys

+++ src/FTPLooper.h
@@ -258,7 +258,7 @@ private:
 	BMessageRunner *fPollingTimer;
 	
 	BString fHost;
-	uint16 fPort;
+	uint32 fPort;
 	BString fUsername;
 	BString fPassword;
 	TDirentParser *fDirentParser;
